# feat: Add Pytest CI mode to optimise migrations