--- a/api/.env-ci
+++ b/api/.env-ci
@@ -1,4 +1,3 @@
 DATABASE_URL=postgresql://postgres:postgres@localhost:5432/postgres
 ANALYTICS_DATABASE_URL=postgres://postgres:postgres@localhost:5432/analytics
-PYTEST_ADDOPTS=--cov . --cov-report xml -n auto --dist worksteal
-GUNICORN_LOGGER_CLASS=util.logging.GunicornJsonCapableLogger
+PYTEST_ADDOPTS=--cov . --cov-report xml -n auto --dist worksteal --ci