--- a/api/.env-local
+++ b/api/.env-local
@@ -1,4 +1,3 @@
 DATABASE_URL=postgresql://postgres:password@localhost:5432/flagsmith
 DJANGO_SETTINGS_MODULE=app.settings.local
 PYTEST_ADDOPTS=--cov . --cov-report html -n auto
-GUNICORN_LOGGER_CLASS=util.logging.GunicornJsonCapableLogger