Thanks for submitting a PR! Please check the boxes below:

- [x] I have run [`pre-commit`](https://docs.flagsmith.com/platform/contributing#pre-commit) to check linting
- [ ] I have added information to `docs/` if required so people know about the feature!
- [x] I have filled in the "Changes" section below?
- [x] I have filled in the "How did you test this code" section below?
- [x] I have used a [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) title for this Pull Request

## Changes

This PR introduces a "CI mode" to Core API test runner, invoked by providing the `--ci` option to pytest.

In this mode, the following optimisations are applied to how pytest manages Django databases when in xdist mode:
 - Databases for workers are created using Django's native cloning method. This means that the migrations are run only once, and all the subsequent databases are copied using `CREATE DATABASE ... WITH TEMPLATE` query.
 - Database setup is only run on the controller node outside of the test session.
 - Database teardown is completely skipped.

Locally, this yields a substantial decrease in overall running time:

 ```sh
make test  708.02s user 34.22s system 216% cpu 5:43.19 total
make test opts='--ci'  558.61s user 29.81s system 255% cpu 3:50.14 total
```

I suspect it won't be as prominent in CI as it probably improves linearly with the number of workers (which is going to be less in CI).

## How did you test this code?

The only proper way was to run the whole test suite.